+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- AdvancedMarkerElement to set some markers on the map. ([#1](https://github.com/Wild-Siena/google-maps-bundle/issues/1))
+- New models LatLng, Marker. ([#1](https://github.com/Wild-Siena/google-maps-bundle/issues/1))
+
+### Changed
+
+- Models GoogleMap, LoaderOptions, MapOptions. ([#1](https://github.com/Wild-Siena/google-maps-bundle/issues/1))
+
+### Removed
+
+### Fixed
+
 ## [0.1.0] 2024-07-05
 
 ### Added